@mladjo2505 2016-07-22
- [ ] prettify pull request description with markdown
- [ ] clean up code from benchmark var dumps after review